adjust styling, rename VServerSettings to ServerSettings
